--- src/types/remark-smartypants.d.ts
@@ -1,5 +1,0 @@
-declare module "@silvenon/remark-smartypants" {
-  import { Plugin } from "unified"
-  const smartypants: Plugin
-  export default smartypants
-}
